RUMM-1401 Fix benchmarks compilation
